change for a Dollar game.
Create a change-counting game that gets the user to enter the number of coins required
to make exactly one dollar. The program should ask the user to enter the number of
pennies, nickels, dimes, and quarters. If the total value of the coins entered is equal to
one dollar, the program should congratulate the user for winning the game. Otherwise,
the program should display a message indicating whether the amount entered was more
than or less than one dollar.